+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - The AccountCapabilities class
 - The MailCapabilities class
+- Refactored Account capabilities to match the spec. #13
 
 ## [0.0.13] - 2016-02-24
 ### Fixed