+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [master]
+
+## [0.0.14] - 2016-05-19
 ### Added
 - The AccountCapabilities class
 - The MailCapabilities class
@@ -10,6 +12,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Account.hasMail
 - Account.hasCalendars
 - Account.hasContacts
+
+### Fixed
+- Message.replyTo is now an array. #39
 
 ## [0.0.13] - 2016-02-24
 ### Fixed